refactor: png 확장자 이미지를 webp 확장자로 수정 후 적용 #300
