+++ frontend/src/components/common/ErrorFallback/AsyncErrorFallback/AsyncErrorFallback.tsx
@@ -8,7 +8,7 @@ import {
   fallbackButtonContainer,
 } from '../ErrorFallback.styled';
 
-import ErrorDdangkong from '@/assets/images/errorDdangkong.png';
+import ErrorDdangkong from '@/assets/images/errorDdangkong.webp';
 import { CustomError } from '@/utils/error';
 
 interface AsyncErrorFallback {
