+++ frontend/src/components/common/ErrorFallback/RouterErrorFallback/RouterErrorFallback.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import Button from '../../Button/Button';
 import { errorFallbackLayout, errorImage, errorText } from '../ErrorFallback.styled';
 
-import ErrorDdangkong from '@/assets/images/errorDdangkong.png';
+import ErrorDdangkong from '@/assets/images/errorDdangkong.webp';
 
 const RouterErrorFallback = () => {
   const navigate = useNavigate();
